# collection.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "postman_collection",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/hotels/{hotel_id}": {
      "get": {
        "summary": "Get Hotel By ID",
        "description": "This endpoint retrieves information about a specific hotel with the ID of 1.\n\n**Genereated using PostbotAI\n\n### Request\n\n- Method: GET\n    \n- URL: `http://localhost:3000/api/hotels/1`\n    \n\n### Response\n\nThe response will include the details of the hotel, such as its name, address, contact information, and any other relevant information.",
        "operationId": "getHotelById",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "summary": "Delete Hotel",
        "description": "### Delete Hotel\n\n**Genereated using PostbotAI\n\nThis endpoint is used to delete a specific hotel by its ID.\n\n#### Request\n\n- Method: DELETE\n    \n- Endpoint: `{{baseURL}}/api/hotels/5`\n    \n\n#### Response\n\nThe response for this request will be a JSON schema representing the success or failure of the deletion operation.",
        "operationId": "deleteHotel",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update Hotel",
        "description": "This endpoint makes an HTTP PATCH request to update the details of a specific hotel with ID 1. The request should include a JSON payload with the updated name, address, location, star rating, and images of the hotel.\n\n**Genereated using PostbotAI\n\n### Request Body\n\n- `name` (string): The updated name of the hotel.\n    \n- `address` (string): The new address of the hotel.\n    \n- `location` (string): The updated location of the hotel.\n    \n- `starRating` (number): The new star rating of the hotel.\n    \n- `images` (array of strings): An array of updated images for the hotel.\n    \n\n### Response\n\nThe response will contain the updated details of the hotel after the PATCH request is processed.",
        "operationId": "updateHotel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "789 New Address"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "updatedImage.png"
                    },
                    "example": [
                      "updatedImage.png"
                    ]
                  },
                  "location": {
                    "type": "string",
                    "example": "Vancouver"
                  },
                  "name": {
                    "type": "string",
                    "example": "Updated Hotel Name333"
                  },
                  "starRating": {
                    "type": "number",
                    "example": 4
                  }
                }
              },
              "examples": {
                "Update Hotel": {
                  "value": {
                    "address": "789 New Address",
                    "images": [
                      "updatedImage.png"
                    ],
                    "location": "Vancouver",
                    "name": "Updated Hotel Name333",
                    "starRating": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "name": "hotel_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/hotels/{hotel_id}/room-types": {
      "post": {
        "summary": "Create Room-Type",
        "description": "### Create Room Type\n\nThis endpoint allows the client to create a new room type for a specific hotel.\n\n**Genereated using PostbotAI\n\n#### Request Body\n\n- `name` (string): The name of the room type.\n    \n- `pricePerNight` (number): The price per night for the room type.\n    \n- `availableRooms` (number): The number of available rooms for the room type.\n    \n- `amenities` (array of strings): The amenities available in the room type.\n    \n- `images` (array of strings): The images of the room type.\n    \n\n#### Response\n\nThe response is in JSON format and includes the schema for the `roomType` object.\n\n``` json\n{\n  \"type\": \"object\",\n  \"properties\": {\n    \"roomType\": {\n      \"type\": \"object\",\n      \"properties\": {\n        \"id\": {\n          \"type\": \"number\"\n        },\n        \"name\": {\n          \"type\": \"string\"\n        },\n        \"amenities\": {\n          \"type\": \"array\",\n          \"items\": {\n            \"type\": \"string\"\n          }\n        },\n        \"pricePerNight\": {\n          \"type\": \"number\"\n        },\n        \"availableRooms\": {\n          \"type\": \"number\"\n        },\n        \"images\": {\n          \"type\": \"array\",\n          \"items\": {\n            \"type\": \"string\"\n          }\n        },\n        \"hotelId\": {\n          \"type\": \"number\"\n        }\n      }\n    }\n  }\n}\n\n ```",
        "operationId": "createRoomType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amenities": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "WiFi"
                    },
                    "example": [
                      "WiFi",
                      "TV",
                      "Mini Bar"
                    ]
                  },
                  "availableRooms": {
                    "type": "number",
                    "example": 5
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "room1.jpg"
                    },
                    "example": [
                      "room1.jpg",
                      "room2.png"
                    ]
                  },
                  "name": {
                    "type": "string",
                    "example": "Deluxe Suite"
                  },
                  "pricePerNight": {
                    "type": "number",
                    "example": 200
                  }
                }
              },
              "examples": {
                "Create Room-Type": {
                  "value": {
                    "amenities": [
                      "WiFi",
                      "TV",
                      "Mini Bar"
                    ],
                    "availableRooms": 5,
                    "images": [
                      "room1.jpg",
                      "room2.png"
                    ],
                    "name": "Deluxe Suite",
                    "pricePerNight": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "name": "hotel_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}